content_factory/requirements.txt
google-generativeai>=0.8.0
aiogram>=3.0.0
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
schedule>=1.2.0
Pillow>=10.0.0
